Renamed files and changed styles for pages
