--- components/Header.module.css
@@ -1,16 +1,0 @@
-.headerContainer {
-	display: flex;
-	justify-content: space-between;
-	margin: 1rem;
-}
-
-.button {
-	background-color: #fdd017;
-	border: none;
-	color: #09215f;
-	padding: 0.5rem;
-	width: 5rem;
-	height: 2rem;
-	border-radius: 5px;
-	/* cursor: pointer; */
-}
