+++ components/Header/Header.module.css
@@ -1,0 +1,16 @@
+.headerContainer {
+    display: flex;
+    justify-content: space-between;
+    margin: 1rem;
+}
+
+.button {
+    background-color: #fdd017;
+    border: none;
+    color: #09215f;
+    margin: 0.5rem;
+    width: 5rem;
+    height: 2rem;
+    border-radius: 5px;
+    cursor: pointer;
+}
