+++ components/Header/Index.js
@@ -1,9 +1,9 @@
-import logo from '../public/images/logo.svg';
+import logo from '../../public/images/logo.svg';
 import styles from './Header.module.css';
 import Image from 'next/image';
 import Link from 'next/link';
 import Head from 'next/head';
-import { supabase } from '../api';
+import { supabase } from '../../api';
 
 export default function Header() {
 	return (
