+++ components/ImpactAreas/ImpactAreas.module.css
@@ -1,5 +1,5 @@
 .main {
-    padding: 0rem 1rem;
+    padding: 1rem;
 }
 
 .grid {
