--- styles/Survey.module.css
@@ -1,65 +1,0 @@
-.container {
-	min-height: 100vh;
-	display: flex;
-	flex-direction: column;
-	height: 100vh;
-}
-
-.main {
-	padding: 3rem 1rem;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	width: 100%;
-}
-
-.card {
-	max-width: 700px;
-}
-
-.input {
-	display: flex;
-	justify-content: center;
-}
-
-.title {
-	margin-bottom: 2rem;
-	font-size: 3rem;
-	text-align: center;
-}
-
-.button {
-	margin: 2rem;
-	background-color: #fdd017;
-	border: none;
-	color: #09215f;
-	padding: 0.5rem;
-	width: 5rem;
-	border-radius: 5px;
-	cursor: pointer;
-}
-
-@media (max-width: 600px) {
-	.grid {
-		width: 100%;
-		flex-direction: column;
-	}
-}
-
-.containerLogin {
-	min-width: 100px;
-	max-width: 500px;
-	margin: 0 auto;
-}
-
-.headerContainer {
-	margin: 1rem;
-	display: flex;
-	justify-content: center;
-	margin-bottom: 5rem;
-}
-
-.login {
-	margin: 1.5rem;
-}
